abstract label sugar to a function
