+++ src/Core/components/Line.js
@@ -11,6 +11,7 @@ import {
   returnRoundingAttributes,
   returnRoundingStateVariableDefinitions,
 } from "../utils/rounding";
+import { returnWrapNonLabelsSugarFunction } from "../utils/label";
 
 export default class Line extends GraphicalComponent {
   constructor(args) {
@@ -67,71 +68,12 @@ export default class Line extends GraphicalComponent {
   static returnSugarInstructions() {
     let sugarInstructions = super.returnSugarInstructions();
 
-    let nonLabelToEquationAttribute = function ({
-      matchedChildren,
-      componentInfoObjects,
-    }) {
-      if (matchedChildren.length === 0) {
-        return { success: false };
-      }
-
-      // wrap first group of non-label children becomes equation
-
-      let componentIsLabel = (x) =>
-        componentInfoObjects.componentIsSpecifiedType(x, "label");
-      let childIsLabel = matchedChildren.map(componentIsLabel);
-
-      let childrenToWrap = [],
-        childrenToNotWrap = [];
-
-      if (childIsLabel.filter((x) => x).length === 0) {
-        childrenToWrap = matchedChildren;
-      } else {
-        if (childIsLabel[0]) {
-          // started with label, find first non-label child
-          let firstNonLabelInd = childIsLabel.indexOf(false);
-          if (firstNonLabelInd !== -1) {
-            childrenToNotWrap.push(
-              ...matchedChildren.slice(0, firstNonLabelInd),
-            );
-            matchedChildren = matchedChildren.slice(firstNonLabelInd);
-            childIsLabel = childIsLabel.slice(firstNonLabelInd);
-          }
-        }
-
-        // now we don't have label at the beginning
-        // find first label ind
-        let firstLabelInd = childIsLabel.indexOf(true);
-        if (firstLabelInd === -1) {
-          childrenToWrap = matchedChildren;
-        } else {
-          childrenToWrap = matchedChildren.slice(0, firstLabelInd);
-          childrenToNotWrap.push(...matchedChildren.slice(firstLabelInd));
-        }
-      }
-
-      if (childrenToWrap.length === 0) {
-        return { success: false };
-      }
-
-      return {
-        success: true,
-        newAttributes: {
-          equation: {
-            component: {
-              componentType: "math",
-              children: childrenToWrap,
-            },
-          },
-        },
-        newChildren: childrenToNotWrap,
-      };
-    };
-
     sugarInstructions.push({
-      replacementFunction: nonLabelToEquationAttribute,
+      replacementFunction: returnWrapNonLabelsSugarFunction({
+        wrappingComponentType: "math",
+        createAttributeOfType: "equation",
+      }),
     });
-
     return sugarInstructions;
   }
