+++ src/Core/utils/label.js
@@ -1,3 +1,109 @@
+export function returnWrapNonLabelsSugarFunction({
+  wrappingComponentType,
+  createAttributeOfType,
+  onlyStringOrMacros = false,
+  customWrappingFunction,
+}) {
+  return function ({ matchedChildren, componentInfoObjects }) {
+    if (matchedChildren.length === 0) {
+      return { success: false };
+    }
+
+    let componentIsLabel = (x) =>
+      componentInfoObjects.componentIsSpecifiedType(x, "label");
+
+    if (
+      onlyStringOrMacros &&
+      !matchedChildren.every(
+        (child) =>
+          typeof child === "string" ||
+          child.doenetAttributes?.createdFromMacro ||
+          componentIsLabel(child),
+      )
+    ) {
+      return { success: false };
+    }
+
+    // wrap first group of non-label children in wrappingComponentType
+
+    let childIsLabel = matchedChildren.map(componentIsLabel);
+
+    let childrenToWrap = [],
+      childrenToNotWrapBegin = [],
+      childrenToNotWrapEnd = [];
+
+    if (childIsLabel.filter((x) => x).length === 0) {
+      childrenToWrap = matchedChildren;
+    } else {
+      if (childIsLabel[0]) {
+        // started with label, find first non-label child
+        let firstNonLabelInd = childIsLabel.indexOf(false);
+        if (firstNonLabelInd !== -1) {
+          childrenToNotWrapBegin = matchedChildren.slice(0, firstNonLabelInd);
+          matchedChildren = matchedChildren.slice(firstNonLabelInd);
+          childIsLabel = childIsLabel.slice(firstNonLabelInd);
+        }
+      }
+
+      // now we don't have label at the beginning
+      // find first label ind
+      let firstLabelInd = childIsLabel.indexOf(true);
+      if (firstLabelInd === -1) {
+        childrenToWrap = matchedChildren;
+      } else {
+        childrenToWrap = matchedChildren.slice(0, firstLabelInd);
+        childrenToNotWrapEnd = matchedChildren.slice(firstLabelInd);
+      }
+    }
+
+    if (childrenToWrap.length === 0) {
+      return { success: false };
+    }
+
+    if (createAttributeOfType) {
+      return {
+        success: true,
+        newAttributes: {
+          [createAttributeOfType]: {
+            component: {
+              componentType: wrappingComponentType,
+              children: childrenToWrap,
+            },
+          },
+        },
+        newChildren: [...childrenToNotWrapBegin, ...childrenToNotWrapEnd],
+      };
+    } else {
+      // apply only if have a single string or multiple children to wrap
+      if (
+        (childrenToWrap.length === 1 &&
+          typeof childrenToWrap[0] !== "string") ||
+        childrenToWrap.length === 0
+      ) {
+        return { success: false };
+      }
+
+      let wrappedChildren;
+      if (customWrappingFunction) {
+        wrappedChildren = customWrappingFunction(childrenToWrap);
+      } else {
+        wrappedChildren = [
+          { componentType: wrappingComponentType, children: childrenToWrap },
+        ];
+      }
+
+      return {
+        success: true,
+        newChildren: [
+          ...childrenToNotWrapBegin,
+          ...wrappedChildren,
+          ...childrenToNotWrapEnd,
+        ],
+      };
+    }
+  };
+}
+
 export function returnLabelStateVariableDefinitions() {
   let stateVariableDefinitions = {};
 
@@ -60,7 +166,7 @@ export function returnLabelStateVariableDefinitions() {
       labelChild: {
         dependencyType: "child",
         childGroups: ["labels"],
-        variableNames: ["value", "hasLatex"],
+        variableNames: ["value", "hasLatex", "hidden"],
       },
       // Note: assuming component has a labelIsName attribute
       // that creates an attribute component and state variable
@@ -86,12 +192,21 @@ export function returnLabelStateVariableDefinitions() {
         dependencyValues.labelChild[dependencyValues.labelChild.length - 1];
 
       if (labelChild && !labelChild.shadowDepth) {
-        return {
-          setValue: {
-            label: labelChild.stateValues.value,
-            labelHasLatex: labelChild.stateValues.hasLatex,
-          },
-        };
+        if (labelChild.stateValues.hidden) {
+          return {
+            setValue: {
+              label: "",
+              labelHasLatex: false,
+            },
+          };
+        } else {
+          return {
+            setValue: {
+              label: labelChild.stateValues.value,
+              labelHasLatex: labelChild.stateValues.hasLatex,
+            },
+          };
+        }
       } else if (essentialValues.label !== undefined) {
         return {
           useEssentialOrDefaultValue: {
